demo/requirements.txt
gradio_gradio_browser_agent_console